+++ cli/src/template.rs
@@ -86,7 +86,7 @@ anchor-lang = "{2}"
 pub fn deploy_js_script_host(cluster_url: &str, script_path: &str) -> String {
     format!(
         r#"
-const anchor = require('@project-serum/anchor');
+const anchor = require('@coral-xyz/anchor');
 
 // Deploy script defined by the user.
 const userScript = require("{0}");
@@ -113,7 +113,7 @@ main();
 
 pub fn deploy_ts_script_host(cluster_url: &str, script_path: &str) -> String {
     format!(
-        r#"import * as anchor from '@project-serum/anchor';
+        r#"import * as anchor from '@coral-xyz/anchor';
 
 // Deploy script defined by the user.
 const userScript = require("{0}");
@@ -143,7 +143,7 @@ pub fn deploy_script() -> &'static str {
 // single deploy script that's invoked from the CLI, injecting a provider
 // configured from the workspace's Anchor.toml.
 
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 
 module.exports = async function (provider) {
   // Configure client to use the provider.
@@ -159,7 +159,7 @@ pub fn ts_deploy_script() -> &'static str {
 // single deploy script that's invoked from the CLI, injecting a provider
 // configured from the workspace's Anchor.toml.
 
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 
 module.exports = async function (provider) {
   // Configure client to use the provider.
@@ -201,7 +201,7 @@ pub struct Initialize {{}}
 
 pub fn mocha(name: &str) -> String {
     format!(
-        r#"const anchor = require("@project-serum/anchor");
+        r#"const anchor = require("@coral-xyz/anchor");
 
 describe("{}", () => {{
   // Configure the client to use the local cluster.
@@ -228,7 +228,7 @@ pub fn package_json() -> String {
         "lint": "prettier */*.js \"*/**/*{{.js,.ts}}\" --check"
     }},
     "dependencies": {{
-        "@project-serum/anchor": "^{0}"
+        "@coral-xyz/anchor": "^{0}"
     }},
     "devDependencies": {{
         "chai": "^4.3.4",
@@ -249,7 +249,7 @@ pub fn ts_package_json() -> String {
         "lint": "prettier */*.js \"*/**/*{{.js,.ts}}\" --check"
     }},
     "dependencies": {{
-        "@project-serum/anchor": "^{0}"
+        "@coral-xyz/anchor": "^{0}"
     }},
     "devDependencies": {{
         "chai": "^4.3.4",
@@ -269,8 +269,8 @@ pub fn ts_package_json() -> String {
 
 pub fn ts_mocha(name: &str) -> String {
     format!(
-        r#"import * as anchor from "@project-serum/anchor";
-import {{ Program }} from "@project-serum/anchor";
+        r#"import * as anchor from "@coral-xyz/anchor";
+import {{ Program }} from "@coral-xyz/anchor";
 import {{ {} }} from "../target/types/{}";
 
 describe("{}", () => {{
@@ -338,7 +338,7 @@ pub fn node_shell(
 ) -> Result<String> {
     let mut eval_string = format!(
         r#"
-const anchor = require('@project-serum/anchor');
+const anchor = require('@coral-xyz/anchor');
 const web3 = anchor.web3;
 const PublicKey = anchor.web3.PublicKey;
 const Keypair = anchor.web3.Keypair;