+++ examples/tutorial/basic-0/tests/basic-0.js
@@ -1,4 +1,4 @@
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 
 describe("basic-0", () => {
   // Configure the client to use the local cluster.