+++ examples/tutorial/basic-1/tests/basic-1.js
@@ -1,5 +1,5 @@
 const assert = require("assert");
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 const { SystemProgram } = anchor.web3;
 
 describe("basic-1", () => {