+++ tests/ido-pool/tests/utils/index.js
@@ -1,5 +1,5 @@
 const spl = require("@solana/spl-token");
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 const serumCmn = require("@project-serum/common");
 const TokenInstructions = require("@project-serum/serum").TokenInstructions;