+++ tests/multiple-suites/tests/fourth-and-fifth-suite/fifth-suite/fifthSuite.ts
@@ -1,5 +1,5 @@
-import * as anchor from "@project-serum/anchor";
-import { Program } from "@project-serum/anchor";
+import * as anchor from "@coral-xyz/anchor";
+import { Program } from "@coral-xyz/anchor";
 import { PublicKey } from "@solana/web3.js";
 import { assert } from "chai";
 import { MultipleSuites } from "../../../target/types/multiple_suites";