+++ tests/spl/token-proxy/tests/token-proxy.js
@@ -1,4 +1,4 @@
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 const { assert } = require("chai");
 
 describe("token", () => {